+++ apps/shell/messages/tr.json
@@ -1,0 +1,3 @@
+{
+  "hello-world": "Merhaba Dünya!"
+}
