+++ apps/shell/src/app/[locale]/governance/proposal/[id]/page.tsx
@@ -8,7 +8,7 @@ import { notFound } from 'next/navigation';
 import { createCosmosService, getChainParams } from '@haqq/data-access-cosmos';
 import { ProposalDetailsPage } from '@haqq/shell-governance';
 import { parseWagmiCookies } from '@haqq/shell-shared';
-import { supportedChainsIds } from '../../../../config/wagmi-config';
+import { supportedChainsIds } from '../../../../../config/wagmi-config';
 
 export const dynamic = 'force-dynamic';
 export const fetchCache = 'force-no-store';
