+++ apps/shell/src/app/[locale]/staking/liquid-staking/@modals/[...catchAll]/page.tsx
