+++ apps/shell/src/app/[locale]/staking/liquid-staking/[...catchAll]/page.tsx
