+++ apps/shell/src/app/[locale]/staking/liquid-staking/page.tsx
