+++ apps/shell/src/app/[locale]/staking/validator/[address]/@modals/[...catchAll]/page.tsx
