+++ apps/shell/src/app/[locale]/staking/validator/[address]/@modals/delegate/page.tsx
