+++ apps/shell/src/app/[locale]/staking/validator/[address]/@modals/redelegate/page.tsx
