+++ apps/shell/src/app/[locale]/staking/validator/[address]/@modals/undelegate/page.tsx
