+++ apps/shell/src/app/[locale]/staking/validator/[address]/[...catchAll]/page.tsx
