+++ apps/shell/src/app/[locale]/staking/validator/[address]/page.tsx
@@ -1,22 +1,29 @@
 import {
-  dehydrate,
   HydrationBoundary,
   QueryClient,
+  dehydrate,
 } from '@tanstack/react-query';
 import { headers } from 'next/headers';
+import { notFound } from 'next/navigation';
 import { createCosmosService, getChainParams } from '@haqq/data-access-cosmos';
-import { MainPage } from '@haqq/shell-main';
 import {
   ethToHaqq,
-  getShellChainStatsData,
   indexerBalancesFetcher,
   parseWagmiCookies,
 } from '@haqq/shell-shared';
-import { supportedChainsIds } from '../config/wagmi-config';
+import { ValidatorDetailsPage } from '@haqq/shell-staking';
+import { supportedChainsIds } from '../../../../../config/wagmi-config';
 
-export default async function IndexPage() {
-  const headersList = headers();
-  const cookies = headersList.get('cookie');
+export default async function ValidatorDetails({
+  params: { address },
+}: {
+  params: { address: string };
+}) {
+  if (!address) {
+    return notFound();
+  }
+
+  const cookies = headers().get('cookie');
   const { chainId, walletAddress } = parseWagmiCookies(cookies);
   const chainIdToUse =
     chainId && supportedChainsIds.includes(chainId)
@@ -25,52 +32,20 @@ export default async function IndexPage() {
 
   const { cosmosRestEndpoint } = getChainParams(chainIdToUse);
   const {
-    getProposals,
-    getGovernanceParams,
     getValidators,
+    getValidatorInfo,
     getStakingPool,
     getStakingParams,
     getRewardsInfo,
     getAccountDelegations,
+    getUndelegations,
   } = createCosmosService(cosmosRestEndpoint);
   const queryClient = new QueryClient();
-  const userAgent = headersList.get('user-agent');
-  const isMobileUserAgent = Boolean(
-    userAgent?.match(
-      /Android|BlackBerry|iPhone|iPad|iPod|Opera Mini|IEMobile|WPDesktop/i,
-    ),
-  );
 
   await queryClient.prefetchQuery({
-    queryKey: [chainIdToUse, 'chain-stats'],
-    queryFn: getShellChainStatsData,
-  });
-
-  await queryClient.prefetchQuery({
-    queryKey: [chainIdToUse, 'proposals'],
-    queryFn: getProposals,
-  });
-
-  await queryClient.prefetchQuery({
-    queryKey: [chainIdToUse, 'governance-params'],
+    queryKey: [chainIdToUse, 'validators'],
     queryFn: async () => {
-      const [deposit_params, voting_params, tally_params] = await Promise.all([
-        getGovernanceParams('deposit').then((res) => {
-          return res.deposit_params;
-        }),
-        getGovernanceParams('voting').then((res) => {
-          return res.voting_params;
-        }),
-        getGovernanceParams('tallying').then((res) => {
-          return res.tally_params;
-        }),
-      ]);
-
-      return {
-        deposit_params,
-        voting_params,
-        tally_params,
-      };
+      return await getValidators();
     },
   });
 
@@ -85,9 +60,9 @@ export default async function IndexPage() {
   });
 
   await queryClient.prefetchQuery({
-    queryKey: [chainIdToUse, 'validators'],
-    queryFn: async () => {
-      return await getValidators();
+    queryKey: [chainIdToUse, 'validator', address],
+    queryFn: () => {
+      return getValidatorInfo(address);
     },
   });
 
@@ -103,12 +78,7 @@ export default async function IndexPage() {
 
     await queryClient.prefetchQuery({
       queryKey: [chainIdToUse, 'rewards', haqqAddress],
-
       queryFn: async () => {
-        if (!haqqAddress) {
-          return null;
-        }
-
         return await getRewardsInfo(haqqAddress);
       },
     });
@@ -116,11 +86,14 @@ export default async function IndexPage() {
     await queryClient.prefetchQuery({
       queryKey: [chainIdToUse, 'delegation', haqqAddress],
       queryFn: async () => {
-        if (!haqqAddress) {
-          return null;
-        }
-
         return await getAccountDelegations(haqqAddress);
+      },
+    });
+
+    await queryClient.prefetchQuery({
+      queryKey: [chainIdToUse, 'unbondings', haqqAddress],
+      queryFn: async () => {
+        return await getUndelegations(haqqAddress);
       },
     });
   }
@@ -129,10 +102,7 @@ export default async function IndexPage() {
 
   return (
     <HydrationBoundary state={dehydratedState}>
-      <MainPage
-        isMobileUserAgent={isMobileUserAgent}
-        seedPhrase={Date.now().toString()}
-      />
+      <ValidatorDetailsPage address={address} />
     </HydrationBoundary>
   );
 }
